[#12] Refactor: HistoryStackView 제거
- 커스텀 클래스이지만 사실상 일반적인 UIStackView 사용과 차이가 없기 때문에 지거
- SideViewTableViewCell 내부의 속성은 UIStackView 타입으로 대체
- 스택뷰에 들어갈 UILabel 속성은 HistoryStackView에서 해당 커스텀 셀 클래스의 속성으로 이동
